wip: consolidate adapters
