+++ src/kairos/workspace/adapter/http/create-workspace/create-workspace-response.ts
